refactor: hover, components
